added authentication & authorization for updating/removing a bug

+++ api/bug/bug.controller.js
@@ -1,6 +1,7 @@
 import PDFDocument from 'pdfkit'
 
 import { bugService } from "./bug.service.js"
+import { authService } from '../auth/auth.service.js'
 
 export async function getBugById(req, res) { 
     try{
@@ -27,11 +28,13 @@ export async function getBugById(req, res) {
 
 export async function saveBug(req, res){
     try {
+        const loginToken = req.cookies.loginToken
+        const user = authService.validateToken(loginToken)
         const bugToSave  = req.body
-        if (bugToSave) {
-            const bug = await bugService.save(bugToSave)
+        if (bugToSave && user.username === bugToSave.creator.username) {
+            const bug = await bugService.save(bugToSave, user)
             res.send(bug)
-        }
+        } else throw 'Saving failed'
     } catch(e) {
         console.log('error in server: ', e)
         res.status(400).send(`Saving failed`)
@@ -76,7 +79,11 @@ export async function getPDF(req, res){
 export async function removeBug(req, res){
     try{
         const { bugId } = req.params
-        await bugService.remove(bugId) 
+        const loginToken = req.cookies.loginToken
+        const user = authService.validateToken(loginToken)
+        const { creator } = bugService.getById(bugId)
+        if (user.username === creator.username ) await bugService.remove(bugId) 
+        else throw 'Removing bug failed'
         res.send(`Removed ID ${bugId} succsesfully`)
     } catch(e) {
         console.log('error in server: ', e)
